+++ src/simplifier/EstablishIntervals.h
@@ -524,13 +524,13 @@ namespace BEEV
           // Ignores what's already there for now..
 
           IntervalType * circ_result = freshUnsignedInterval(n.GetValueWidth());
-          for (int i=0; i < n[0].GetValueWidth()-1;i++)
+          for (int i=0; i < (int)n[0].GetValueWidth()-1;i++)
           {
               CONSTANTBV::BitVector_Bit_On(circ_result->maxV,i);
               CONSTANTBV::BitVector_Bit_Off(circ_result->minV,i);
           }
 
-          for (int i=n[0].GetValueWidth()-1; i < n.GetValueWidth();i++)
+          for (int i = (int)n[0].GetValueWidth()-1; i < n.GetValueWidth();i++)
           {
               CONSTANTBV::BitVector_Bit_Off(circ_result->maxV,i);
               CONSTANTBV::BitVector_Bit_On(circ_result->minV,i);
@@ -605,7 +605,7 @@ namespace BEEV
           CONSTANTBV::BitVector_increment(result->maxV);
 
           bool bad= false;
-          for (int i =0; i < children.size(); i++)
+          for (size_t i =0; i < children.size(); i++)
             {
               if (children[i] == NULL)
                 {
@@ -621,7 +621,7 @@ namespace BEEV
               if (CONSTANTBV::Set_Max(max) >= width)
                 bad = true;
 
-              for (int j = width; j < 2 * width; j++)
+              for (unsigned j = width; j < 2 * width; j++)
                 {
                   if (CONSTANTBV::BitVector_bit_test(min, j))
                     bad = true;
@@ -723,7 +723,7 @@ namespace BEEV
           if (knownC1)
           {
         	  // Copy in the minimum and maximum.
-        	  for (int i=0; i < n[1].GetValueWidth();i++)
+        	  for (unsigned i=0; i < n[1].GetValueWidth();i++)
         	  {
         		  if (CONSTANTBV::BitVector_bit_test(children[1]->maxV,i))
         			  CONSTANTBV::BitVector_Bit_On(result->maxV,i);
@@ -803,10 +803,10 @@ namespace BEEV
 
     ~EstablishIntervals()
     {
-      for (int i =0; i < toDeleteLater.size();i++)
+      for (size_t i =0; i < toDeleteLater.size();i++)
         delete toDeleteLater[i];
 
-      for (int i =0; i < likeAutoPtr.size();i++)
+      for (size_t i =0; i < likeAutoPtr.size();i++)
         CONSTANTBV::BitVector_Destroy(likeAutoPtr[i]);
 
       likeAutoPtr.clear();
